requirements.txt
Pillow~=10.0.1
reportlab~=4.0.6
future~=0.18.3
ttkthemes~=3.2.2
frontend~=0.0.3
tools~=0.1.9
PyMuPDF~=1.24.0